+++ libs/cardano-ledger-core/src/Cardano/Ledger/Block.hs
@@ -23,12 +23,19 @@ module Cardano.Ledger.Block (
   bodyBlockL,
   bheader,
   bbody,
+  neededTxInsForBlock,
   BbodySignal (..),
   EraBlockHeader (..),
-  neededTxInsForBlock,
+  EbAnnouncement (..),
 ) where
 
 import Cardano.Ledger.BaseTypes (ProtVer)
+import Cardano.Ledger.Binary (
+  DecCBOR (decCBOR),
+  EncCBOR (..),
+  decodeRecordNamed,
+  encodeListLen,
+ )
 import Cardano.Ledger.Core
 import Cardano.Ledger.TxIn (TxIn (..))
 import Cardano.Slotting.Slot (SlotNo)
@@ -110,3 +117,26 @@ class Era era => EraBlockHeader h era where
   blockBodyHashBlockHeaderL :: Lens' (Block h era) (Hash HASH EraIndependentBlockBody)
   slotNoBlockHeaderL :: Lens' (Block h era) SlotNo
   protVerBlockHeaderL :: Lens' (Block h era) ProtVer
+
+-- | Announcement of an Endorser Block (EB).
+data EbAnnouncement = EbAnnouncement
+  { ebAnnouncementHash :: !(SafeHash EraIndependentEb)
+  -- ^ Hash of the announced Endorsement Block References
+  , ebAnnouncementSize :: !Word32
+  -- ^ Size of the announced Endorsement Block References
+  }
+  deriving stock (Show, Eq, Generic)
+  deriving anyclass (NoThunks, NFData)
+
+instance EncCBOR EbAnnouncement where
+  encCBOR (EbAnnouncement h s) =
+    encodeListLen 2
+      <> encCBOR h
+      <> encCBOR s
+
+instance DecCBOR EbAnnouncement where
+  decCBOR =
+    decodeRecordNamed "EbAnnouncement" (const 2) $
+      EbAnnouncement
+        <$> decCBOR
+        <*> decCBOR