+++ libs/cardano-protocol/testlib/Test/Cardano/Protocol/Leios/BlockHeader/Arbitrary.hs
@@ -14,11 +14,10 @@ import qualified Cardano.Crypto.KES as KES
 import Cardano.Crypto.Util (SignableRepresentation)
 import qualified Cardano.Crypto.VRF as VRF
 import Cardano.Ledger.Binary (DecCBOR)
-import Cardano.Ledger.Block (Block (Block))
+import Cardano.Ledger.Block (Block (Block), EbAnnouncement (EbAnnouncement))
 import Cardano.Ledger.Core (BlockBody, EraBlockBody)
 import Cardano.Protocol.Crypto (Crypto (KES, VRF))
 import Cardano.Protocol.Leios.BlockHeader (
-  EbAnnouncement (EbAnnouncement),
   Header (Header, HeaderConstr),
   HeaderBody (HeaderBody),
  )